Switched Manifest extension to ManifestLoader and sank it into PackageLoading.
