+++ Examples/package-info/Sources/package-info/main.swift
@@ -30,7 +30,7 @@ let package = localFileSystem.currentWorkingDirectory!
 
 // There are several levels of information available.
 // Each takes longer to load than the level above it, but provides more detail.
-let manifest: Manifest = try Manifest.loadManifest(from: package, with: swiftCompiler)
+let manifest: Manifest = try ManifestLoader.loadManifest(from: package, with: swiftCompiler)
 let loadedPackage: Package = try Package.loadPackage(from: package, with: swiftCompiler)
 let graph: PackageGraph = try PackageGraph.loadGraph(from: package, with: swiftCompiler)
