+++ Sources/PackageLoading/ManifestLoader.swift
@@ -180,6 +180,25 @@ public final class ManifestLoader: ManifestLoaderProtocol {
        )
     }
 
+    /// Loads a manifest from a package repository using the resources associated with a particular `swift` executable.
+    ///
+    /// - Parameters:
+    ///     - package: The absolute path of the package root.
+    ///     - swiftExecutable: The absolute path of a `swift` executable.
+    ///         Its associated resources will be used by the loader.
+    public static func loadManifest(
+        from package: AbsolutePath,
+        with swiftExecutable: AbsolutePath) throws -> Manifest {
+
+        let resources = try UserManifestResources(swiftExectuable: swiftExecutable)
+        let loader = ManifestLoader(manifestResources: resources)
+        let toolsVersion = try ToolsVersionLoader().load(at: package, fileSystem: localFileSystem)
+        return try loader.load(
+            package: package,
+            baseURL: package.pathString,
+            manifestVersion: toolsVersion.manifestVersion)
+    }
+
     public func load(
         packagePath path: AbsolutePath,
         baseURL: String,
